Crate ast

This crate explores genetic programming. Genetic programming is

a technique whereby computer programs are encoded as a set of genes that are then modified (evolved) using an evolutionary algorithm.

The problem we are trying to tackle with genetic programming is that of landing a moon lander. I.e. land a moon lander safely on the surface of a moon without user intervention. A nice interactive game that gives a feel for the problem can be found here.

Modules

darwin

data
The data model is responsible for creation and evaluation of data in the moon lander domain.

random
The random module creates random ast::structure::Program.

simulation

source
The source module provides a representation of ast::structure::Programs

structure
The structure is the abstract syntax tree for the moon lander control code.

Macros

F!

T!

and!

constant!

divide!

equal!

fuel!

greater!

greater_equal!

iff!

left!

less!

less_equal!

minus!

multiply!

not!

o!

or!

pick!
The pick macro allows to randomly select one expression from a sequence of expressions according to associated weights.

plus!

right!

skip!

thrust!

vx!

vy!

w!

x!

y!