Module ast::structure

The structure is the abstract syntax tree for the moon lander control code.

It is the hearth of the crate. Creating a Program by hand is a little bit tedious. The tedium is a little bit relieved with the introduction of macros, the possibility to read from and write to JSON and to generate a random program.

Enums

Command
The Command enum is used as an argument to ast::structure::Program::Command

Condition
The Condition enum is used as the condition in a ast::Program::If

Expression
The Expression enum is used as the comparative ast::structure::Conditions

Program
The Program enum is the root of the abstract syntax tree.

Sensor
The Sensor enum is used in ast::Structure::Expression as input to calculations.