Enum ast::structure::Command

pub enum Command {
    Skip,
    Left,
    Right,
    Thrust,
}

The Command enum is used as an argument to ast::structure::Program::Command

You can find syntactic sugar to create Commands more easily in macros

Examples

extern crate ast;

let command: ast::structure::Command = ast::structure::Command::Skip;

Variants

Skip

Left

Right

Thrust