Enum ast::structure::Condition

pub enum Condition {
    True,
    False,
    Not(Box<Condition>),
    Or(Box<Condition>, Box<Condition>),
    And(Box<Condition>, Box<Condition>),
    Less(Box<Expression>, Box<Expression>),
    LessEqual(Box<Expression>, Box<Expression>),
    Equal(Box<Expression>, Box<Expression>),
    GreaterEqual(Box<Expression>, Box<Expression>),
    Greater(Box<Expression>, Box<Expression>),
}

The Condition enum is used as the condition in a ast::Program::If

The following groups are present in Condition: literal: it includes Condition::True and Condition::False. logical: it includes Condition::Not, Condition::Or and Condition::And. comparative: it includes Condition::Less, Condition::LessEqual, Condition::Equal, Condition::GreaterEqual, and Condition::Greater

You can find syntactic sugar to create Conditions more easily in macros

Examples

extern crate ast;

let condition: ast::structure::Condition = ast::structure::Condition::And(
  Box::new(ast::structure::Condition::Not(Box::new(ast::structure::Condition::False))),
  Box::new(ast::structure::Condition::Less(
    Box::new(ast::structure::Expression::Constant(1.0)),
    Box::new(ast::structure::Expression::Constant(2.0))
  ))
);

Variants
TrueFalseNot(Box<Condition>)Or(Box<Condition>, Box<Condition>)And(Box<Condition>, Box<Condition>)Less(Box<Expression>, Box<Expression>)LessEqual(Box<Expression>, Box<Expression>)Equal(Box<Expression>, Box<Expression>)GreaterEqual(Box<Expression>, Box<Expression>)Greater(Box<Expression>, Box<Expression>)