Enum ast::structure::Program

pub enum Program {
    If(Box<Condition>, Box<Program>, Box<Program>),
    Command(Box<Command>),
}

The Program enum is the root of the abstract syntax tree.

Program is either a single Command or an If-statement.

You can find syntactic sugar to create Programs more easily in macros.

Examples

extern crate ast;

let program: ast::structure::Program = ast::structure::Program::If(
  Box::new(ast::structure::Condition::True),
  Box::new(ast::structure::Program::Command(Box::new(ast::structure::Command::Skip))),
  Box::new(ast::structure::Program::Command(Box::new(ast::structure::Command::Thrust)))
);

Variants

If

Command