Enum ast::structure::Expression

pub enum Expression {
    Constant(f32),
    Sensor(Box<Sensor>),
    Plus(Box<Expression>, Box<Expression>),
    Minus(Box<Expression>, Box<Expression>),
    Multiply(Box<Expression>, Box<Expression>),
    Divide(Box<Expression>, Box<Expression>),
}

The Expression enum is used as the comparative ast::structure::Conditions

It allows you to do calculations with Expression::Constants or Expression::Sensors

You can find syntactic sugar to create Expressions more easily in macros

Examples

extern crate ast;

let expression: ast::structure::Expression = ast::structure::Expression::Plus(
  Box::new(ast::structure::Expression::Constant(42.0)),
  Box::new(ast::structure::Expression::Sensor(Box::new(ast::structure::Sensor::Vx)))
);

Variants

Constant

Sensor

Plus

Minus

Multiply

Divide